Add files via upload
I added some comments, and made some adjustments. My suggestions/additions are highlighted in green. Rather than erasing anyone's work, I simply lined through what was written if I think it could be removed (much of it was just leftover from the template). If it's highlighted in yellow and I didn't make any changes to it, that means that I agree with what is written. Let me know what else needs to be done, or what I can do to help. The doc looks good so far.